requirements.txt
requests
pytesseract
Pillow
beautifulsoup4
PyPDF2
docx2txt
openpyxl
pywin32